Don't use ports and adapters for ImportGraph
It's never brought any value to have an abstraction for the
ImportGraph. So this removes the abstract class and then moves the
concretion to the application layer.
