+++ .importlinter
@@ -18,6 +18,6 @@ ignore_imports =
     ; is an object within grimp, rather than a module in its own right,
     ; so we ignore these imports here.
     grimp.adaptors.caching -> grimp
-    grimp.adaptors.graph -> grimp
+    grimp.application.graph -> grimp
     grimp.adaptors.filesystem -> grimp
     grimp.application.scanning -> grimp
