+++ src/grimp/main.py
@@ -2,7 +2,7 @@ __all__ = ["build_graph"]
 
 from .adaptors.caching import Cache
 from .adaptors.filesystem import FileSystem
-from .adaptors.graph import ImportGraph
+from .application.graph import ImportGraph
 from .adaptors.modulefinder import ModuleFinder
 from .adaptors.packagefinder import ImportLibPackageFinder
 from .adaptors.timing import SystemClockTimer
