+++ tests/unit/application/graph/test_direct_imports.py
@@ -1,6 +1,6 @@
 import pytest  # type: ignore
 
-from grimp.adaptors.graph import ImportGraph
+from grimp.application.graph import ImportGraph
 import re
 from grimp.exceptions import InvalidImportExpression
