+++ tests/unit/application/graph/test_hierarchy.py
@@ -1,6 +1,6 @@
 import pytest  # type: ignore
 
-from grimp.adaptors.graph import ImportGraph
+from grimp.application.graph import ImportGraph
 from grimp.exceptions import ModuleNotPresent, InvalidModuleExpression
