+++ tests/unit/application/graph/test_misc.py
@@ -1,6 +1,6 @@
 import pytest  # type: ignore
 
-from grimp.adaptors.graph import ImportGraph
+from grimp.application.graph import ImportGraph
 from grimp.exceptions import ModuleNotPresent
 
 
@@ -27,12 +27,3 @@ class TestIsModuleSquashed:
 
         with pytest.raises(ModuleNotPresent):
             assert not graph.is_module_squashed("foo")
-
-
-class TestFindAllSimpleChains:
-    def test_removed_exception(self):
-        with pytest.raises(
-            AttributeError,
-            match="This method has been removed. Consider using find_shortest_chains instead?",
-        ):
-            ImportGraph().find_all_simple_chains(importer="foo", imported="bar")
