+++ tests/unit/application/graph/test_repr.py
@@ -1,6 +1,6 @@
 import re
 
-from grimp.adaptors.graph import ImportGraph
+from grimp.application.graph import ImportGraph
 
 
 class TestRepr:
